STIS-69:  Renamed routines to hide internal code (#97)
* STIS-69:  Changed relative imports to hide internal code.

* Fix to mkfringeflat doc parameter list for Sphinx.
