+++ stistools/defringe/_findloc.py
