+++ stistools/defringe/_fit1d.py
